+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 -   updated to bevy 0.15
 -   changed plugin name to `AsepriteUltraPlugin`.
 -   removed bundles, switched to required components.
+-   added `ManualTick` component. Let's you update the animation state following you own logic.
+-   added `FrameChangedEvent`. Triggering it on an entity ensures a frame re-render. (has to be called manual if in manual control mode).
 -   added `thiserror`.
 -   removed `anyhow`.